reorganize packages for service test fixtures
