+++ core/api/endpoint/site/src/test/java/org/example/age/api/endpoint/site/SiteApiEndpointTest.java
@@ -15,7 +15,7 @@ import org.example.age.data.crypto.AesGcmEncryptionPackage;
 import org.example.age.data.crypto.DigitalSignature;
 import org.example.age.data.crypto.SecureId;
 import org.example.age.data.user.VerifiedUser;
-import org.example.age.service.endpoint.site.stub.StubSiteComponent;
+import org.example.age.service.component.stub.avs.StubSiteComponent;
 import org.example.age.testing.client.TestClient;
 import org.example.age.testing.server.TestServer;
 import org.example.age.testing.server.undertow.TestUndertowServer;
