+++ core/integration-test/src/test/java/org/example/age/service/endpoint/ServiceEndpointTest.java
@@ -1,8 +1,8 @@
 package org.example.age.service.endpoint;
 
 import java.io.IOException;
-import org.example.age.service.endpoint.avs.test.TestAvsComponent;
-import org.example.age.service.endpoint.site.test.TestSiteComponent;
+import org.example.age.service.component.test.avs.TestAvsComponent;
+import org.example.age.service.component.test.site.TestSiteComponent;
 import org.example.age.testing.server.TestServer;
 import org.example.age.testing.server.undertow.TestUndertowServer;
 import org.example.age.testing.service.ServiceIntegrationTestTemplate;
