+++ core/service/endpoint/avs/src/test/java/org/example/age/service/endpoint/avs/AvsServiceEndpointTest.java
@@ -1,8 +1,8 @@
 package org.example.age.service.endpoint.avs;
 
 import java.io.IOException;
-import org.example.age.service.endpoint.avs.test.TestAvsComponent;
-import org.example.age.service.endpoint.site.fake.FakeSiteComponent;
+import org.example.age.service.component.fake.site.FakeSiteComponent;
+import org.example.age.service.component.test.avs.TestAvsComponent;
 import org.example.age.testing.server.TestServer;
 import org.example.age.testing.server.undertow.TestUndertowServer;
 import org.example.age.testing.service.ServiceIntegrationTestTemplate;
