+++ core/service/endpoint/avs/src/test/java/org/example/age/service/verification/internal/avs/AvsVerificationManagerTest.java
@@ -16,9 +16,9 @@ import org.example.age.data.crypto.SecureId;
 import org.example.age.data.user.AgeRange;
 import org.example.age.data.user.VerifiedUser;
 import org.example.age.module.extractor.common.builtin.DisabledAuthMatchData;
-import org.example.age.service.verification.internal.avs.test.TestAvsVerificationComponent;
+import org.example.age.service.component.fake.site.FakeSiteVerificationComponent;
+import org.example.age.service.component.test.avs.TestAvsVerificationComponent;
 import org.example.age.service.verification.internal.site.FakeSiteVerificationProcessor;
-import org.example.age.service.verification.internal.site.fake.FakeSiteVerificationComponent;
 import org.example.age.testing.api.StubScheduledExecutor;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
