+++ core/service/endpoint/avs/src/testFixtures/java/org/example/age/service/component/fake/site/TestFakeSiteVerificationProcessorModule.java
@@ -1,0 +1,16 @@
+package org.example.age.service.component.fake.site;
+
+import dagger.Module;
+import org.example.age.module.key.common.test.TestKeyModule;
+import org.example.age.module.store.common.inmemory.InMemoryVerificationStoreModule;
+import org.example.age.service.verification.internal.site.FakeSiteVerificationProcessor;
+import org.example.age.service.verification.internal.site.FakeSiteVerificationProcessorModule;
+
+/** Dagger module that binds dependencies for {@link FakeSiteVerificationProcessor}. */
+@Module(
+        includes = {
+            FakeSiteVerificationProcessorModule.class,
+            InMemoryVerificationStoreModule.class,
+            TestKeyModule.class,
+        })
+interface TestFakeSiteVerificationProcessorModule {}
