+++ core/service/endpoint/site/src/test/java/org/example/age/service/verification/internal/site/SiteVerificationManagerTest.java
@@ -15,9 +15,9 @@ import org.example.age.data.crypto.AesGcmEncryptionPackage;
 import org.example.age.data.crypto.DigitalSignature;
 import org.example.age.data.user.VerifiedUser;
 import org.example.age.module.extractor.common.builtin.UserAgentAuthMatchData;
+import org.example.age.service.component.fake.avs.FakeAvsVerificationComponent;
+import org.example.age.service.component.test.site.TestSiteVerificationComponent;
 import org.example.age.service.verification.internal.avs.FakeAvsVerificationFactory;
-import org.example.age.service.verification.internal.avs.fake.FakeAvsVerificationComponent;
-import org.example.age.service.verification.internal.site.test.TestSiteVerificationComponent;
 import org.example.age.testing.api.StubScheduledExecutor;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
