--- core/service/endpoint/site/src/testFixtures/java/org/example/age/service/verification/internal/avs/fake/FakeAvsVerificationComponent.java
@@ -1,26 +1,0 @@
-package org.example.age.service.verification.internal.avs.fake;
-
-import dagger.Component;
-import javax.inject.Singleton;
-import org.example.age.service.verification.internal.avs.FakeAvsVerificationFactory;
-
-/** Component that creates a test {@link FakeAvsVerificationFactory}. */
-public final class FakeAvsVerificationComponent {
-
-    /** Creates a {@link FakeAvsVerificationFactory}. */
-    public static FakeAvsVerificationFactory createFakeAvsVerificationFactory() {
-        UnderlyingComponent component = DaggerFakeAvsVerificationComponent_UnderlyingComponent.create();
-        return component.fakeAvsVerificationFactory();
-    }
-
-    // static class
-    private FakeAvsVerificationComponent() {}
-
-    /** Dagger component that provides a {@link FakeAvsVerificationFactory}. */
-    @Component(modules = TestFakeAvsVerificationFactoryModule.class)
-    @Singleton
-    interface UnderlyingComponent {
-
-        FakeAvsVerificationFactory fakeAvsVerificationFactory();
-    }
-}
